# src/AdaSoftLibrary.Api/Swagger/AdaSoftLibrary.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AdaSoftLibrary Web API",
    "description": "Version: 1.0.0.0, Date: 11. 7. 2024 17:23:04",
    "version": "v1"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "AdaSoftLibrary.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "summary": "Web API Info",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/books": {
      "get": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Zoznam kníh",
        "parameters": [
          {
            "name": "BookFilter",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetBookResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Zaevidovanie novej knihy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Detail knihy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Editácia údajov o knihe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Vymazanie knihy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/books/{id}/borrow": {
      "post": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Vypožičanie knihy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/books/{id}/return": {
      "post": {
        "tags": [
          "BookEndpoint"
        ],
        "summary": "Vrátenie knihy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "HomeEndpoint"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthenticationRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Prihlásovacie údaje používateľa"
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BorrowBookRequest": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Vypožičanie knihy"
      },
      "CreateBookRequest": {
        "type": "object",
        "properties": {
          "author": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Zaevidovanie novej knihy"
      },
      "GetBookResponse": {
        "required": [
          "author",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "minLength": 1,
            "type": "string",
            "description": "Autor"
          },
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Názov"
          },
          "year": {
            "type": "integer",
            "description": "Rok",
            "format": "int32",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Popis",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Meno",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Priezvisko",
            "nullable": true
          },
          "reader": {
            "type": "string",
            "description": "Meno čitateľa",
            "nullable": true,
            "readOnly": true
          },
          "borrowedFrom": {
            "type": "string",
            "description": "Dátum požičania",
            "format": "date",
            "nullable": true
          },
          "isBorrowed": {
            "type": "boolean",
            "description": "Dostupnosť"
          }
        },
        "additionalProperties": false,
        "description": "Detail knihy"
      },
      "UpdateBookRequest": {
        "type": "object",
        "properties": {
          "author": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Editácia údajov o knihe"
      }
    },
    "securitySchemes": {
      "AdaSoftLibrary BasicAuth": {
        "type": "http",
        "description": "Basic HTTP authentication",
        "scheme": "Basic"
      }
    }
  },
  "security": [
    {
      "AdaSoftLibrary BasicAuth": [ ]
    }
  ]
}